+++ src/WebJobs.Script.Rpc/IWorkerChannel.cs
