+++ src/WebJobs.Script.Rpc/ProcessManagement/DefaultWorkerProcessFactory.cs
