+++ src/WebJobs.Script.Rpc/ProcessManagement/EmptyProcessRegistry.cs
