+++ src/WebJobs.Script.Rpc/ProcessManagement/IProcessRegistry.cs
