+++ src/WebJobs.Script.Rpc/ProcessManagement/IWorkerProcess.cs
