+++ src/WebJobs.Script.Rpc/ProcessManagement/IWorkerProcessFactory.cs
