+++ src/WebJobs.Script.Rpc/ProcessManagement/JobObjectRegistry.cs
