+++ src/WebJobs.Script.Rpc/ProcessManagement/WorkerContext.cs
