+++ src/WebJobs.Script.Rpc/ProcessManagement/WorkerProcess.cs
