+++ src/WebJobs.Script.Rpc/ProcessManagement/WorkerProcessExitException.cs
