+++ src/WebJobs.Script.Rpc/ProcessManagement/WorkerProcessUtilities.cs
