+++ src/WebJobs.Script.Rpc/Rpc/IJobHostRpcWorkerChannelManager.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 using System;
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 
 namespace Microsoft.Azure.WebJobs.Script.Workers.Rpc
 {
-    internal interface IJobHostRpcWorkerChannelManager
+    public interface IJobHostRpcWorkerChannelManager
     {
         void AddChannel(IRpcWorkerChannel channel, string language);
