+++ src/WebJobs.Script.Rpc/WebJobs.Script.Rpc.csproj
@@ -11,4 +11,8 @@
     <ProjectReference Include="..\WebJobs.Script\WebJobs.Script.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Configuration\" />
+  </ItemGroup>
+
 </Project>
