+++ src/WebJobs.Script.Rpc/WorkerUtilities.cs
