+++ src/WebJobs.Script/Workers/WorkerDescription.cs
