+++ src/WebJobs.Script/Workers/WorkerProcessArguments.cs
