+++ src/WebJobs.Script/Workers/WorkerProcessCountOptions.cs
